Reorganized recipe step folders.
