+++ src/Orchard.Tests.Modules/Recipes/RecipeHandlers/ModuleStepTest.cs
@@ -23,7 +23,7 @@ using Orchard.Packaging.GalleryServer;
 using Orchard.Packaging.Models;
 using Orchard.Packaging.Services;
 using Orchard.Recipes.Models;
-using Orchard.Recipes.RecipeExecutionSteps;
+using Orchard.Recipes.Providers.Executors;
 using Orchard.Recipes.Services;
 using Orchard.Tests.Environment.Extensions;
 using Orchard.Tests.Environment.Features;
