+++ src/Orchard.Web/Modules/Orchard.ImportExport/Commands/ImportExportCommands.cs
@@ -4,10 +4,10 @@ using System.IO;
 using System.Linq;
 using Orchard.Commands;
 using Orchard.ContentManagement.MetaData;
-using Orchard.ImportExport.RecipeBuilderSteps;
+using Orchard.ImportExport.Recipes.Builders;
 using Orchard.ImportExport.Services;
 using Orchard.Recipes.Models;
-using Orchard.Recipes.RecipeBuilders;
+using Orchard.Recipes.Providers.Builders;
 using Orchard.Recipes.Services;
 using Orchard.Security;
 using Orchard.Settings;
