+++ src/Orchard.Web/Modules/Orchard.ImportExport/Orchard.ImportExport.csproj
@@ -71,7 +71,7 @@
     <Compile Include="Models\ExportOptions.cs" />
     <Compile Include="Permissions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RecipeBuilderSteps\CustomStepsStep.cs" />
+    <Compile Include="Recipes\Builders\CustomStepsStep.cs" />
     <Compile Include="Services\ExportContext.cs" />
     <Compile Include="ViewModels\CustomStepEntry.cs" />
     <Compile Include="Services\ICustomExportStep.cs" />
