+++ src/Orchard.Web/Modules/Orchard.Recipes/Orchard.Recipes.csproj
@@ -82,16 +82,16 @@
     <Compile Include="Models\RecipeStepResultRecord.cs" />
     <Compile Include="Models\VersionHistoryOptions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RecipeBuilders\ContentStep.cs" />
-    <Compile Include="RecipeBuilders\RecipeMetadataStep.cs" />
-    <Compile Include="RecipeBuilders\SettingsStep.cs" />
-    <Compile Include="RecipeExecutionSteps\CommandStep.cs" />
-    <Compile Include="RecipeExecutionSteps\ContentStep.cs" />
-    <Compile Include="RecipeExecutionSteps\ContentSchemaStep.cs" />
-    <Compile Include="RecipeExecutionSteps\MigrationStep.cs" />
-    <Compile Include="RecipeExecutionSteps\ModuleStep.cs" />
-    <Compile Include="RecipeExecutionSteps\SettingsStep.cs" />
-    <Compile Include="RecipeHandlers\RecipeExecutionStepHandler.cs" />
+    <Compile Include="Providers\Builders\ContentStep.cs" />
+    <Compile Include="Providers\Builders\RecipeMetadataStep.cs" />
+    <Compile Include="Providers\Builders\SettingsStep.cs" />
+    <Compile Include="Providers\Executors\CommandStep.cs" />
+    <Compile Include="Providers\Executors\ContentStep.cs" />
+    <Compile Include="Providers\Executors\ContentSchemaStep.cs" />
+    <Compile Include="Providers\Executors\MigrationStep.cs" />
+    <Compile Include="Providers\Executors\ModuleStep.cs" />
+    <Compile Include="Providers\Executors\SettingsStep.cs" />
+    <Compile Include="Providers\RecipeHandlers\RecipeExecutionStepHandler.cs" />
     <Compile Include="Routes.cs" />
     <Compile Include="Services\RecipeHarvester.cs" />
     <Compile Include="Services\RecipeManager.cs" />
