+++ src/Orchard.Web/Modules/Orchard.Recipes/Providers/Executors/MigrationStep.cs
@@ -5,7 +5,7 @@ using Orchard.Data.Migration;
 using Orchard.Logging;
 using Orchard.Recipes.Services;
 
-namespace Orchard.Recipes.RecipeExecutionSteps {
+namespace Orchard.Recipes.Providers.Executors {
     public class MigrationStep : RecipeExecutionStep {
         private readonly IDataMigrationManager _dataMigrationManager;
